move voting to the current branch
