+++ scripts/vote_2023_06_20_2.py
