+++ tests/test_2023_06_20_2.py
@@ -1,8 +1,8 @@
 """
-Tests for voting
+Tests for voting 2023_06_20_2
 
 """
-from scripts.vote_stake_eth import start_vote
+from scripts.vote_2023_06_20_2 import start_vote
 
 from utils.config import (
     network_name,
